Tiddied up the directory by moving files over.
The directory kinda loked messy, and seeing as i couldn't use subdirs
for files (It's a wierd work around to get python to import files in
dirs). I moved them into a new directory.

This also means you just have to download the "Program" folder instead
of the whole repo to get the program to run.