Removed Json from newSave (below), Fix place issue
So, reason for removing json was because it was making binary format
unsecure. Even though i encode the json object using pickle, Apparently
for some reason the data would still be shown in the file if you did
something as simple as 'less FileName'. Because of this (and to keep
data secure), i had to remove it. Yes, the data is now easier to read
and write, but most people aren't going to bother so it's fine.